refactor: extract print from StatePrinter.hpp
